Bug fixes:     -Fixed a bug in copy/past operation     -Fixed icone display by windows explorer for ffd files     -Fixed display of waveforme in edit video dialog (remove extra space to the right)     -Fixed a display bug of translation in about dialog on Windows     -Improve audio/video synchronization
New features:
    -Also saving the position of the vertical scrollbar in the edit slides dialog
    -Upgrade ffmpeg on Windows to version 2.2.1

+++ trunk/ffDiaporama/src/ffDiaporama/engine/cBaseMediaFile.h
@@ -502,7 +502,7 @@ private:
     };
 private slots:
     void DecodeAudio(sAudioContext *AudioContext,AVPacket *StreamPacket,int64_t Position);
-    void DecodeAudioFrame(sAudioContext *AudioContext,int64_t *FramePts,AVFrame *Frame,int64_t Position);
+    void DecodeAudioFrame(sAudioContext *AudioContext,qreal *FramePts,AVFrame *Frame,int64_t Position);
 };
 
 //*********************************************************************************************************************************************
